refactor(event-requests): extract lifecycle routes from legacy router
